Heavily changed how our database is generated and laid out.
Changed from using the built-in rails id column to using our own twitter_uid column as the primary key for the users table. Since this column needs to hold 64bit values, this required us to swap to using sql instead of rails to store our persistent database schema. Due to the extensive scaffolding changes, we've also reset the database migrations, so a manual database drop/create is required if you update to this version.