We now download and store the latest 200 tweets from the user's timeline in our DB.
Had to add the foreigner gem to do database-level foreign keys.
Added the tweet model that handles tweet creation and storage. It also makes sure the user that made the tweet is in our database.
The user model has an update_tweets method that queries twitter for that user's tweets.