Changed our database structure slightly.
We now use the activerecord-native_db_types_override gem to set id columns as bigints, this allows us to simplify our migrations going forward. Please note that all data is lost when executing this migration. Also note that we've gone back to using the built-in ruby database definition format instead of raw SQL that we were forced to use before.